# docs/swagger.yaml
definitions:
...
  reimbursement.ApprovalInfoResponse:
    properties:
      action:
        type: string
      action_at:
        type: string
      approver_name:
        type: string
      comment:
        type: string
      id:
        type: integer
    type: object
  reimbursement.CreateReimbursementRequest:
    properties:
      department_id:
        description: 部门ID
        type: integer
      employee_id:
        description: 工号
        type: string
      employee_name:
        description: 申请人姓名
        type: string
      submit_note:
        description: 报销事由
        type: string
    required:
    - department_id
    - employee_id
    - employee_name
    type: object
  reimbursement.InvoiceItemResponse:
    properties:
      amount:
        type: integer
      category:
        type: string
      check_result:
        type: string
      id:
        type: integer
      invoice_date:
        type: string
    type: object
  reimbursement.ListReimbursementResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/reimbursement.ReimbursementResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  reimbursement.ReimbursementResponse:
    properties:
      approvals:
        items:
          $ref: '#/definitions/reimbursement.ApprovalInfoResponse'
        type: array
      created_at:
        type: string
      department:
        type: string
      department_id:
        type: integer
      employee_id:
        type: string
      employee_name:
        type: string
      id:
        type: integer
      invoices:
        items:
          $ref: '#/definitions/reimbursement.InvoiceItemResponse'
        type: array
      need_special_approval:
        type: boolean
      reimbursement_no:
        type: string
      status:
        type: string
      submit_note:
        type: string
      total_amount:
        type: integer
      updated_at:
        type: string
    type: object
  reimbursement.SubmitReimbursementRequest:
    properties:
      total_amount:
        description: 报销总金额(分)
        type: integer
    required:
    - total_amount
    type: object
  reimbursement.UploadInvoiceResponse:
    properties:
      file_id:
        description: 文件唯一标识（UUID）
        type: string
      file_name:
        description: 原始文件名
        type: string
      file_path:
        description: 存储路径（供 Agent OCR 工具使用）
        type: string
      size:
        description: 文件大小（字节）
        type: integer
      url:
        description: 可访问 URL（用于前端预览）
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/reimbursements:
    get:
      consumes:
      - application/json
      description: 分页查询报销单，可按员工工号筛选
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认10
        in: query
        name: page_size
        type: integer
      - description: 员工工号（可选筛选）
        in: query
        name: employee_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 报销单列表
          schema:
            $ref: '#/definitions/reimbursement.ListReimbursementResponse'
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取报销单列表
      tags:
      - 报销管理
    post:
      consumes:
      - application/json
      description: 员工创建报销单草稿，状态为 draft，需后续提交
      parameters:
      - description: 创建报销单请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reimbursement.CreateReimbursementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 报销单创建成功
          schema:
            $ref: '#/definitions/reimbursement.ReimbursementResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建报销单（草稿）
      tags:
      - 报销管理
  /api/reimbursements/{id}:
    get:
      consumes:
      - application/json
      description: 根据报销单ID获取报销单详细信息，包含票据和审批记录
      parameters:
      - description: 报销单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 报销单详情
          schema:
            $ref: '#/definitions/reimbursement.ReimbursementResponse'
        "400":
          description: 报销单ID格式错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 报销单不存在
          schema:
            additionalProperties: true
            type: object
      summary: 获取报销单详情（按ID）
      tags:
      - 报销管理
...
  /api/reimbursements/{id}/approve:
    post:
      consumes:
      - application/json
      description: 审批人通过报销单，更新报销单状态并记录审批操作
      parameters:
      - description: 报销单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 审批通过成功
          schema:
            $ref: '#/definitions/reimbursement.ReimbursementResponse'
        "400":
          description: 报销单ID格式错误
          schema:
            additionalProperties: true
            type: object
        "409":
          description: 审批操作失败（状态不允许或无权限）
          schema:
            additionalProperties: true
            type: object
      summary: 审批通过报销单
      tags:
      - 报销管理
  /api/reimbursements/{id}/reject:
    post:
      consumes:
      - application/json
      description: 审批人驳回报销单，报销单状态回退为 draft
      parameters:
      - description: 报销单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 驳回成功
          schema:
            $ref: '#/definitions/reimbursement.ReimbursementResponse'
        "400":
          description: 报销单ID格式错误
          schema:
            additionalProperties: true
            type: object
        "409":
          description: 驳回操作失败（状态不允许或无权限）
          schema:
            additionalProperties: true
            type: object
      summary: 驳回报销单
      tags:
      - 报销管理
  /api/reimbursements/{id}/submit:
    post:
      consumes:
      - application/json
      description: 将草稿状态的报销单提交进入审批流程
      parameters:
      - description: 报销单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 提交报销单请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reimbursement.SubmitReimbursementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 报销单提交成功
          schema:
            $ref: '#/definitions/reimbursement.ReimbursementResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "409":
          description: 提交失败（状态不允许或预算不足）
          schema:
            additionalProperties: true
            type: object
      summary: 提交报销单
      tags:
      - 报销管理
  /api/reimbursements/no/{no}:
    get:
      consumes:
      - application/json
      description: 根据报销单号获取报销单详细信息，包含票据和审批记录
      parameters:
      - description: 报销单号（如 REIMB-2026-0001）
        in: path
        name: "no"
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 报销单详情
          schema:
            $ref: '#/definitions/reimbursement.ReimbursementResponse'
        "404":
          description: 报销单不存在
          schema:
            additionalProperties: true
            type: object
      summary: 获取报销单详情（按单号）
      tags:
      - 报销管理
  /api/reimbursements/pending:
    get:
      consumes:
      - application/json
      description: 获取所有状态为待审批的报销单列表（审批人专用）
      produces:
      - application/json
      responses:
        "200":
          description: 待审批报销单列表
          schema:
            items:
              $ref: '#/definitions/reimbursement.ReimbursementResponse'
            type: array
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取待审批报销单
      tags:
      - 报销管理
  /api/reimbursements/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传票据图片文件（支持 JPG/PNG/PDF），保存到文件存储（本地磁盘或 MinIO），返回文件路径供 Agent OCR
        工具使用。
      parameters:
      - description: 票据图片文件（JPG/PNG/PDF/BMP/TIFF，最大 10MB）
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功，返回文件路径和访问 URL
          schema:
            $ref: '#/definitions/reimbursement.UploadInvoiceResponse'
        "400":
          description: 未选择文件或文件类型不支持
          schema:
            additionalProperties: true
            type: object
        "413":
          description: 文件过大
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 文件存储失败
          schema:
            additionalProperties: true
            type: object
      summary: 上传票据图片
      tags:
      - 报销管理
swagger: "2.0"